Move GitHubItem in interface package Implement interface in GitHubReposResponse and GitHubUsersResponse classes
