+++ app/src/main/java/peter/staranchuk/githubclient/network/response/GitHubUsersResponse.kt
@@ -1,8 +1,11 @@
 package peter.staranchuk.githubclient.network.response
 
 import com.google.gson.annotations.SerializedName
-import peter.staranchuk.githubclient.model.GitHubItem
+import peter.staranchuk.githubclient.enums.GitHubItemType
+import peter.staranchuk.githubclient.interfaces.GitHubItem
 
 data class GitHubUsersResponse(@SerializedName("items") val users : List<UserInfo>)
 
-data class UserInfo(val login : String, val score : Double, @SerializedName("avatar_url") val avatarUrl : String)  : GitHubItem
+data class UserInfo(val login : String, val score : Double, @SerializedName("avatar_url") val avatarUrl : String)  : GitHubItem {
+    override fun getCode(): Int = GitHubItemType.USER.code
+}
